server/requirements.txt
cheroot==10.0.1
CherryPy==18.10.0
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0